requirements.txt
Django==3.1.4
django-cors-headers==3.5.0
djangorestframework==3.12.2
boto3==1.14.56
botocore==1.17.56
celery==4.4.6
redis==3.5.3
requests==2.25.1
decorator==4.4.2
ipython==7.16.1
ipython-genutils==0.2.0
pytz==2020.4
python-dateutil==2.8.1
mysqlclient==2.0.1
django-environ==0.4.5
psycopg2-binary==2.8.6
pycryptodome==3.10.1
django-redis==4.12.1
